Validation on delete permission

+++ src/backend/ent/src/main/java/com/davidticona/ent/service/impl/PermissionServiceImpl.java
@@ -5,6 +5,7 @@ import com.davidticona.ent.domain.dto.permission.PermissionResponseDto;
 import com.davidticona.ent.util.Tree.TreeNode;
 import com.davidticona.ent.domain.entity.Permission;
 import com.davidticona.ent.domain.repository.PermissionRepository;
+import com.davidticona.ent.exceptions.ConflictException;
 import com.davidticona.ent.service.PermissionService;
 import com.davidticona.ent.util.Tree.AdjacentItem;
 import com.davidticona.ent.util.Tree.Tree;
@@ -12,6 +13,7 @@ import com.davidticona.ent.util.mapper.PermissionMapper;
 import com.davidticona.ent.validator.ObjectValidator;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
+import java.util.LinkedList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -67,6 +69,16 @@ public class PermissionServiceImpl implements PermissionService{
     public void delete(Integer id) {
         Permission permission = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException());
+        List<String> errors = new LinkedList<>();
+        if (hasRoles(id)) {
+            errors.add("Unable to delete record as it has associated roles");
+        }
+        if (hasChildren(id)) {
+            errors.add("Unable to delete record as it has associated permissions");
+        }
+        if (!errors.isEmpty()) {
+            throw new ConflictException(errors);
+        }
         repository.delete(permission);
     }
 
@@ -79,4 +91,14 @@ public class PermissionServiceImpl implements PermissionService{
     public boolean permissionExists(List<Integer> ids) {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }    
+
+    @Override
+    public boolean hasRoles(Integer id) {
+        return (repository.countRoles(id) > 0);
+    }
+
+    @Override
+    public boolean hasChildren(Integer id) {
+        return (repository.countChildren(id) > 0);
+    }
 }
